created 'extras' directory, moved tsx/scss files
