+++ src/App.tsx
@@ -5,15 +5,15 @@ import React, { useEffect, useState, useRef } from "react";
 import "./App.scss";
 import TypingTest, { TypingTestProps } from "./components/TypingTest/TypingTest";
 import TestResults, { TestResultsProps } from "./components/TestResults/TestResults";
-import ResetButton, { ResetButtonProps } from "./components/ResetButton";
+import ResetButton, { ResetButtonProps } from "./components/Extras/ResetButton";
 import TestOptions from "./components/TestOptions/TestOptions";
-import CompletionBar from "./components/CompletionBar";
-import WordsPerMinute, { WordsPerMinuteProps } from "./components/WordsPerMinute";
-import CapsLockIndicator, { CapsLockIndicatorProps } from "./components/CapsLockIndicator";
-import AfkDetectedIndicator, { AfkDetectedIndicatorProps } from "./components/AfkDetectedIndicator";
+import CompletionBar from "./components/Extras/CompletionBar";
+import WordsPerMinute, { WordsPerMinuteProps } from "./components/Extras/WordsPerMinute";
+import CapsLockIndicator, { CapsLockIndicatorProps } from "./components/Extras/CapsLockIndicator";
+import AfkDetectedIndicator, { AfkDetectedIndicatorProps } from "./components/Extras/AfkDetectedIndicator";
 import Footer, { FooterProps } from "./components/Footer";
 import KeyTips from "./components/KeyTips";
-import ColourPaletteSelector, { ColourPaletteSelectorProps } from "./components/ColourPaletteSelector";
+import ColourPaletteSelector, { ColourPaletteSelectorProps } from "./components/Extras/ColourPaletteSelector";
 import UpdateCssVariable from "./components/HelperComponents/UpdateCssVariable";
 import Header, { HeaderProps } from "./components/Header/Header";
 import { useUserContext } from "./contexts/UserContext";
