+++ src/components/Extras/AfkDetectedIndicator.tsx
@@ -1,4 +1,4 @@
-import "../styles/componentStyles/afk-detected-indicator.scss";
+import "./afk-detected-indicator.scss";
 
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
@@ -9,12 +9,11 @@ export interface AfkDetectedIndicatorProps {
 }
 
 export const AfkDetectedIndicator = ({isAfkMidTest}: AfkDetectedIndicatorProps) => {
-
 	return (
 		<div style={{"opacity": isAfkMidTest ? 1 : 0}} className="afk-indicator-container">
 			<div className="afk-indicator-contents">
 				<FontAwesomeIcon icon={faKeyboard} className="afk-icon"/>
-				Afk dectected last test
+				AFK dectected last test
 			</div>
 		</div>
 	);
