+++ src/components/Extras/capslock-indicator.scss
@@ -1,5 +1,4 @@
-@use "../_variables.scss";
-@use "../_constants.scss";
+@use "../../styles/constants";
 
 .capslock-indicator-container {
   opacity: 0;
