+++ src/components/Extras/words-per-minute.scss
@@ -1,5 +1,3 @@
-@use "../_variables.scss";
-
 .WPM-div {
     grid-area: 7/1;
     position: relative;
